requirements.txt
Django==1.10.6
django-appconf==1.0.2
django-bootstrap3==8.2.1
django-filter==1.0.1
djangorestframework==3.5.4
djangorestframework-csv==2.0.0
djangorestframework-gis==0.11
ete2==2.3.10
psycopg2==2.7
python-nexus==1.42
six==1.10.0
wsgiref==0.1.2
coverage==4.3.4
flake8==3.3.0
newick==0.7
clldutils==1.9.4
attrs==16.3.0
pyglottolog==0.3.1